+++ sakura_core/doc/BlockComment.cpp
@@ -79,7 +79,7 @@ bool BlockComment::Match_CommentFrom(
 
 	@return 当てはまった位置を返すが、当てはまらなかったときは、nLineLenをそのまま返す。
 */
-int BlockComment::Match_CommentTo(
+size_t BlockComment::Match_CommentTo(
 	int					nPos,	// [in] 探索開始位置
 	const StringRef&	str		// [in] 探索対象文字列 ※探索開始位置のポインタではないことに注意
 	/*
